PATCH: Refined transaction report generation
and updated email links

This update makes the calculation for loss and overall profit in transaction reports more accurate. The false-positive check for loss value has been removed and an overdueAmount now includes 'DISPUTED' payments as well. Further, corrected the redirect URL for the "Go to the dashboard" button in the shop report email to correctly route to the platform shop reports.