PATCH: Implement device restriction toggling functionality
Added a new feature that allows for toggling the restriction status of a device in the DeviceScreen class. Also introduced the ability to view linked user details by making user names clickable, redirecting to device management routes in the DeviceList layout. Moreover, a toast info message is displayed to indicate a successful update in restriction status.