PATCH: Change product lookup method from EAN to UPC-A
Updated the product retrieval and validation methods in the ShopProducts model to use UPC-A instead of the EAN code. The changes involved updating check digit removal and product ID extraction processes as well as modifying the validation method to fit UPC-A rules. Also, updated the DeviceController to accommodate these changes.