PATCH: Update CSS and JS in Fluffici vendor files
This commit introduces changes to the dashy.css and dashy.js files within the Fluffici vendor directory. In the CSS file, the changes include additional styling for various elements and inclusion of Bootstrap v5.2.3. As for the JS file, there seem to be changes regarding the animation interactions for the UI elements.