Add numeric and icon fields to metrics
Modified the metric layouts and screens by adding 'numeric' and 'icon' properties to various metric items. 'numeric' is used to determine whether a metric's value is numeric or not, and 'icon' is to provide an associated icon. Ensuring that applications handle metric values correctly and provide a more visually appealing and understandable representation to the users.