Move service-worker.js to public directory
The service-worker.js is moved to the public directory in the src folder as part of the refactoring process for Beams integration. Relocating the file improves the organization of the project structure and better accommodates the simplified PusherPushNotifications.Client implementation and the new streamlined process for displaying web push notifications.