Add product management functions to DeviceController
A series of new methods have been added to the DeviceController in order to handle the retrieval, update, and validation of products. These methods include 'products', 'fetchProduct', 'incrementProduct', and 'voucherInfo'. Each of these methods serve a specific purpose in managing product data in the application, improving the efficiency and functionality of the system.