Refactor input components and update barcode generation URL
This commit refactors the structure of the HTML input components in 'input.blade.php' to improve code clarity. The barcode generation URL in 'ShopProductEdit.php' is also updated to use a new endpoint. In addition, redundant RelativeTime component calls in 'AuditLogsListLayout.php' and 'AttachmentsLayout.php' are replaced with Input component calls for displaying date and time.