Add Inventory Management Screens and Layouts
Created 'InventoryEditScreen', 'InventoryList', 'InventoryScreen' and 'StockMovementList' under the Shop module. These are designed to manage the product inventory to track available stocks and product movements effectively. The 'InventoryEditScreen' handles the editing and saving of inventory data while the 'InventoryList', 'InventoryScreen' and 'StockMovementList' present a clear interface for inventory data and movements.